Major Algorithmic Updates and Clean Up
Added a new algorithm to find all beacon matches for a given
advertisement. This is a very clunky solution and should be moved out of
the Node App and onto a server side Java platform. Also cleaned up the
mediabuyer_routes to include tighter error checking. All routes still
need significant amounts of cleanup, and the matches_routes should be
depreciated quickly in favor of a separate server side solution.